Add Backup
Well I had a friend test it and he put it on his desktop so it perma deleted quite a few things and it was not a very good thing so we add backup folder now and move files instead of deleting.  Then we prompt the user if they would like to delete the backup folder at the end of the program.