Reshade Folder & Braap/Soundtable Files
Added ignore to the reshade folder. Removed the depth restriction because I realized if you have a track in a folder like 'demos' or 'setups' you do not deserve that track to be optimized.

Added braap/soundtable file compatibility. It reads in any braap files that might be found in the current directory, and it it finds an associated soundtable it reads in those files to ignore. It will ignore braap files found in subfolders as well.

Added global variable for the backup folder in case I want to change it later.

Added automatic deletion of the backup folder if the cleanup ran deleted 0 items and no files were found in the backup folder.